Created functions in automation and testing
